refactor(backend): restructure survival into a sub-package
Moves the survival analysis code from two flat files (survival.py +
survival_handlers.py) into a proper sub-package with the handler
classes split by comparison family:

  survival/
    __init__.py       (public re-exports — backwards compat)
    router.py         (FastAPI endpoint + _get_endpoint_config)
    handlers/
      __init__.py     (re-exports all handlers + factory)
      base.py         (SurvivalHandler ABC + shared plumbing)
      variant_type.py (VariantTypeHandler)
      pathogenicity.py (PathogenicityHandler)
      disease_subtype.py (DiseaseSubtypeHandler)
      protein_domain.py (ProteinDomainHandler)
      factory.py      (SurvivalHandlerFactory)

Every file is under 280 LOC (was: single 1,055 LOC handlers module
plus a 106 LOC router). Callers using
`from app.phenopackets.routers.aggregations.survival import
SurvivalHandlerFactory` (etc.) continue to work via the __init__.py
re-exports. The one existing test import in
tests/test_survival_protein_domain.py was updated from
`...aggregations.survival_handlers` to `...aggregations.survival`.

One small targeted refactor was required to enable the clean split:
_sql_list was previously a static method on VariantTypeHandler, and
PathogenicityHandler / DiseaseSubtypeHandler / ProteinDomainHandler
all called it via the cross-class reference
`VariantTypeHandler._sql_list(...)`. That's promoted to
`SurvivalHandler._sql_list` on the base class, and all four
subclasses now call `self._sql_list(...)`. Behavior is unchanged;
the cross-class dependency is removed.

pyproject.toml per-file-ignore updated to point at the new glob
`app/phenopackets/routers/aggregations/survival/handlers/*.py` for
E501 (SQL query strings) and D102 (concrete overrides of ABC
methods).

All 812 tests pass (794 baseline + 18 Wave 3 Task 2 smoke tests);
mypy + ruff clean.

Closes P5 #23 from the 2026-04-09 review.
