+++ backend/pyproject.toml
@@ -121,7 +121,7 @@ convention = "google"
 "app/phenopackets/clinical_queries.py" = ["E501"]  # JSONPath and SQL queries
 "app/phenopackets/endpoints.py" = ["E501", "D417"]  # SQL queries and API param descriptions
 "app/phenopackets/routers/aggregations.py" = ["E501"]  # Complex SQL queries with VEP classification
-"app/phenopackets/routers/aggregations/survival_handlers.py" = ["E501", "D102"]  # SQL queries, interface methods
+"app/phenopackets/routers/aggregations/survival/handlers/*.py" = ["E501", "D102"]  # SQL query strings, concrete overrides of ABC methods
 "app/phenopackets/validation/variant_validator.py" = ["E501"]  # Long regex patterns
 "app/phenopackets/variant_search_validation.py" = ["E501"]  # Long comments
 "app/core/patterns.py" = ["E501"]  # Long regex patterns
